# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Celluloid Video Analysis API",
    "version": "1.5.1"
  },
  "paths": {
    "/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Health Check",
        "description": "Health check endpoint",
        "operationId": "health_check",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/job/analyse": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Create an analysis task for a video",
        "description": "Start video analysis on a video",
        "operationId": "create_analysis_job",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/status/{job_id}": {
      "get": {
        "tags": [
          "status"
        ],
        "summary": "Get the status of a detection job",
        "description": "Get the status of a detection job",
        "operationId": "get_job_status",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/job/{job_id}/results": {
      "get": {
        "tags": [
          "results"
        ],
        "summary": "Get the results of a completed analysis job",
        "operationId": "get_job_results",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResultsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "webhooks": {
    "job-completed": {
      "post": {
        "tags": [
          "webhooks"
        ],
        "summary": "Job Completed",
        "description": "When a job is completed, we'll send you a POST request with this\ndata to the URL that you register for the event `job-completed` in the dashboard.",
        "operationId": "job_completedjob_completed_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCompletedWebhook"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AnalysisRequest": {
        "properties": {
          "external_id": {
            "type": "string",
            "title": "External Id",
            "description": "Project identifier"
          },
          "video_url": {
            "type": "string",
            "title": "Video Url",
            "description": "URL or path to video file"
          },
          "similarity_threshold": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Similarity Threshold",
            "description": "Similarity threshold for object tracking",
            "default": 0.5
          },
          "callback_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Callback Url",
            "description": "Callback URL for job completion notifications"
          }
        },
        "type": "object",
        "required": [
          "external_id",
          "video_url"
        ],
        "title": "AnalysisRequest"
      },
      "AnalysisResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "queue_position": {
            "type": "integer",
            "title": "Queue Position"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "callback_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Callback Url"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status",
          "queue_position",
          "message"
        ],
        "title": "AnalysisResponse"
      },
      "BoundingBoxModel": {
        "properties": {
          "x": {
            "type": "integer",
            "title": "X"
          },
          "y": {
            "type": "integer",
            "title": "Y"
          },
          "width": {
            "type": "integer",
            "title": "Width"
          },
          "height": {
            "type": "integer",
            "title": "Height"
          }
        },
        "type": "object",
        "required": [
          "x",
          "y",
          "width",
          "height"
        ],
        "title": "BoundingBoxModel"
      },
      "DetectionFrameModel": {
        "properties": {
          "frame_idx": {
            "type": "integer",
            "title": "Frame Idx"
          },
          "timestamp": {
            "type": "number",
            "title": "Timestamp"
          },
          "objects": {
            "items": {
              "$ref": "#/components/schemas/DetectionObjectModel"
            },
            "type": "array",
            "title": "Objects"
          }
        },
        "type": "object",
        "required": [
          "frame_idx",
          "timestamp",
          "objects"
        ],
        "title": "DetectionFrameModel"
      },
      "DetectionObjectModel": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "class_name": {
            "type": "string",
            "title": "Class Name"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "bbox": {
            "$ref": "#/components/schemas/BoundingBoxModel"
          },
          "thumbnail": {
            "type": "string",
            "title": "Thumbnail"
          }
        },
        "type": "object",
        "required": [
          "id",
          "class_name",
          "confidence",
          "bbox",
          "thumbnail"
        ],
        "title": "DetectionObjectModel"
      },
      "DetectionResultsModel": {
        "properties": {
          "version": {
            "type": "string",
            "title": "Version"
          },
          "metadata": {
            "$ref": "#/components/schemas/ResultsMetadataModel"
          },
          "frames": {
            "items": {
              "$ref": "#/components/schemas/DetectionFrameModel"
            },
            "type": "array",
            "title": "Frames"
          }
        },
        "type": "object",
        "required": [
          "version",
          "metadata",
          "frames"
        ],
        "title": "DetectionResultsModel"
      },
      "DetectionStatisticsModel": {
        "properties": {
          "total_detections": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Detections"
          },
          "person_detections": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Person Detections"
          },
          "person_with_face": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Person With Face"
          },
          "person_without_face": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Person Without Face"
          },
          "other_detections": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Other Detections"
          },
          "class_counts": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "integer"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Class Counts"
          }
        },
        "type": "object",
        "title": "DetectionStatisticsModel"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "version": {
            "type": "string",
            "title": "Version"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          },
          "job_stats": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobStats"
              },
              {
                "type": "null"
              }
            ]
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "version",
          "status",
          "timestamp"
        ],
        "title": "HealthResponse"
      },
      "JobCompletedWebhook": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "external_id": {
            "type": "string",
            "title": "External Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "external_id",
          "status",
          "timestamp"
        ],
        "title": "JobCompletedWebhook"
      },
      "JobResultsResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DetectionResultsModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "JobResultsResponse"
      },
      "JobStats": {
        "properties": {
          "queued": {
            "type": "integer",
            "title": "Queued"
          },
          "processing": {
            "type": "integer",
            "title": "Processing"
          },
          "completed": {
            "type": "integer",
            "title": "Completed"
          },
          "failed": {
            "type": "integer",
            "title": "Failed"
          }
        },
        "type": "object",
        "required": [
          "queued",
          "processing",
          "completed",
          "failed"
        ],
        "title": "JobStats"
      },
      "JobStatusResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "external_id": {
            "type": "string",
            "title": "External Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress": {
            "type": "number",
            "title": "Progress"
          },
          "queue_position": {
            "type": "integer",
            "title": "Queue Position",
            "default": 0
          },
          "estimated_wait_time": {
            "type": "string",
            "title": "Estimated Wait Time",
            "default": "00:00:00"
          },
          "start_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time"
          },
          "end_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "external_id",
          "status",
          "progress"
        ],
        "title": "JobStatusResponse"
      },
      "ProcessingMetadataModel": {
        "properties": {
          "start_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Time"
          },
          "end_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Time"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "frames_processed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frames Processed"
          },
          "frames_with_detections": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frames With Detections"
          },
          "processing_speed": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing Speed"
          },
          "detection_statistics": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DetectionStatisticsModel"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "ProcessingMetadataModel"
      },
      "ResultsMetadataModel": {
        "properties": {
          "video": {
            "$ref": "#/components/schemas/VideoMetadataModel"
          },
          "sprite": {
            "$ref": "#/components/schemas/SpriteMetadataModel"
          },
          "processing": {
            "$ref": "#/components/schemas/ProcessingMetadataModel"
          }
        },
        "type": "object",
        "required": [
          "video",
          "sprite",
          "processing"
        ],
        "title": "ResultsMetadataModel"
      },
      "SpriteMetadataModel": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "thumbnail_size": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Thumbnail Size"
          }
        },
        "type": "object",
        "required": [
          "path",
          "thumbnail_size"
        ],
        "title": "SpriteMetadataModel"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VideoMetadataModel": {
        "properties": {
          "fps": {
            "type": "number",
            "title": "Fps"
          },
          "frame_count": {
            "type": "integer",
            "title": "Frame Count"
          },
          "width": {
            "type": "integer",
            "title": "Width"
          },
          "height": {
            "type": "integer",
            "title": "Height"
          },
          "source": {
            "type": "string",
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "fps",
          "frame_count",
          "width",
          "height",
          "source"
        ],
        "title": "VideoMetadataModel"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  },
  "tags": [
    {
      "name": "analysis",
      "description": "Operations with video analysis."
    },
    {
      "name": "health",
      "description": "Health check operations."
    },
    {
      "name": "status",
      "description": "Status check operations."
    },
    {
      "name": "results",
      "description": "Results operations."
    },
    {
      "name": "webhooks",
      "description": "Webhook operations."
    }
  ]
}